cambio en t_sample_weight
Se cambia el parametro de value int a value float.

+++ pyplexos/solution/zip/xml.py
@@ -244,7 +244,7 @@ class SampleWeightTable(BaseModel):
 
     sample_id: int
     phase_id: int
-    value: int
+    value: float  # antes era int?
 
 
 class CustomCoumnTable(BaseModel):
